ToothFairy/algorithms/Tomasz_Szczepanski/docker/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking
#
arrow==1.2.3
    # via jinja2-time
binaryornot==0.4.4
    # via cookiecutter
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via requests
chardet==5.1.0
    # via binaryornot
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   cookiecutter
    #   evalutils
    #   pip-tools
cookiecutter==2.1.1
    # via evalutils
evalutils==0.4.0
    # via -r requirements.in
idna==3.4
    # via requests
imageio[tifffile]==2.26.0
    # via evalutils
jinja2==3.1.2
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
joblib==1.2.0
    # via scikit-learn
markupsafe==2.1.2
    # via jinja2
monai==1.1.0
numpy==1.24.2
    # via
    #   evalutils
    #   imageio
    #   pandas
    #   scikit-learn
    #   scipy
    #   tifffile
packaging==23.0
    # via build
pandas==1.5.3
    # via evalutils
pillow==9.4.0
    # via imageio
pip-tools==6.12.3
    # via evalutils
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   arrow
    #   pandas
python-slugify==8.0.1
    # via cookiecutter
pytz==2022.7.1
    # via pandas
pyyaml==6.0
    # via cookiecutter
requests==2.28.2
    # via cookiecutter
scikit-learn==1.2.2
    # via evalutils
scipy==1.10.1
    # via
    #   evalutils
    #   scikit-learn
simpleitk==2.2.1
    # via evalutils
six==1.16.0
    # via python-dateutil
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2023.3.15
    # via imageio
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
urllib3==1.26.15
    # via requests
wheel==0.40.0
    # via pip-tools
torchio==0.18.92

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools